Fix bug in xy_candidate_phase_paths
